+++ sdk/resourcemanager/dataprotection/armdataprotection/fetchcrossregionrestorejob_client.go
@@ -20,61 +20,63 @@ import (
 	"strings"
 )
 
-// OperationStatusResourceGroupContextClient contains the methods for the OperationStatusResourceGroupContext group.
-// Don't use this type directly, use NewOperationStatusResourceGroupContextClient() instead.
-type OperationStatusResourceGroupContextClient struct {
+// FetchCrossRegionRestoreJobClient contains the methods for the FetchCrossRegionRestoreJob group.
+// Don't use this type directly, use NewFetchCrossRegionRestoreJobClient() instead.
+type FetchCrossRegionRestoreJobClient struct {
 	internal       *arm.Client
 	subscriptionID string
 }
 
-// NewOperationStatusResourceGroupContextClient creates a new instance of OperationStatusResourceGroupContextClient with the specified values.
+// NewFetchCrossRegionRestoreJobClient creates a new instance of FetchCrossRegionRestoreJobClient with the specified values.
 //   - subscriptionID - The ID of the target subscription. The value must be an UUID.
 //   - credential - used to authorize requests. Usually a credential from azidentity.
 //   - options - pass nil to accept the default values.
-func NewOperationStatusResourceGroupContextClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*OperationStatusResourceGroupContextClient, error) {
+func NewFetchCrossRegionRestoreJobClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*FetchCrossRegionRestoreJobClient, error) {
 	cl, err := arm.NewClient(moduleName, moduleVersion, credential, options)
 	if err != nil {
 		return nil, err
 	}
-	client := &OperationStatusResourceGroupContextClient{
+	client := &FetchCrossRegionRestoreJobClient{
 		subscriptionID: subscriptionID,
 		internal:       cl,
 	}
 	return client, nil
 }
 
-// Get - Gets the operation status for an operation over a ResourceGroup's context.
+// Get - Fetches the Cross Region Restore Job
 // If the operation fails it returns an *azcore.ResponseError type.
 //
 // Generated from API version 2023-11-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
-//   - options - OperationStatusResourceGroupContextClientGetOptions contains the optional parameters for the OperationStatusResourceGroupContextClient.Get
+//   - location - The name of the Azure region.
+//   - parameters - Request body for operation
+//   - options - FetchCrossRegionRestoreJobClientGetOptions contains the optional parameters for the FetchCrossRegionRestoreJobClient.Get
 //     method.
-func (client *OperationStatusResourceGroupContextClient) Get(ctx context.Context, resourceGroupName string, operationID string, options *OperationStatusResourceGroupContextClientGetOptions) (OperationStatusResourceGroupContextClientGetResponse, error) {
+func (client *FetchCrossRegionRestoreJobClient) Get(ctx context.Context, resourceGroupName string, location string, parameters CrossRegionRestoreJobRequest, options *FetchCrossRegionRestoreJobClientGetOptions) (FetchCrossRegionRestoreJobClientGetResponse, error) {
 	var err error
-	const operationName = "OperationStatusResourceGroupContextClient.Get"
+	const operationName = "FetchCrossRegionRestoreJobClient.Get"
 	ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, operationName)
 	ctx, endSpan := runtime.StartSpan(ctx, operationName, client.internal.Tracer(), nil)
 	defer func() { endSpan(err) }()
-	req, err := client.getCreateRequest(ctx, resourceGroupName, operationID, options)
+	req, err := client.getCreateRequest(ctx, resourceGroupName, location, parameters, options)
 	if err != nil {
-		return OperationStatusResourceGroupContextClientGetResponse{}, err
+		return FetchCrossRegionRestoreJobClientGetResponse{}, err
 	}
 	httpResp, err := client.internal.Pipeline().Do(req)
 	if err != nil {
-		return OperationStatusResourceGroupContextClientGetResponse{}, err
+		return FetchCrossRegionRestoreJobClientGetResponse{}, err
 	}
 	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
 		err = runtime.NewResponseError(httpResp)
-		return OperationStatusResourceGroupContextClientGetResponse{}, err
+		return FetchCrossRegionRestoreJobClientGetResponse{}, err
 	}
 	resp, err := client.getHandleResponse(httpResp)
 	return resp, err
 }
 
 // getCreateRequest creates the Get request.
-func (client *OperationStatusResourceGroupContextClient) getCreateRequest(ctx context.Context, resourceGroupName string, operationID string, options *OperationStatusResourceGroupContextClientGetOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DataProtection/operationStatus/{operationId}"
+func (client *FetchCrossRegionRestoreJobClient) getCreateRequest(ctx context.Context, resourceGroupName string, location string, parameters CrossRegionRestoreJobRequest, options *FetchCrossRegionRestoreJobClientGetOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DataProtection/locations/{location}/fetchCrossRegionRestoreJob"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -83,11 +85,11 @@ func (client *OperationStatusResourceGroupContextClient) getCreateRequest(ctx co
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	if operationID == "" {
-		return nil, errors.New("parameter operationID cannot be empty")
+	if location == "" {
+		return nil, errors.New("parameter location cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{operationId}", url.PathEscape(operationID))
-	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	urlPath = strings.ReplaceAll(urlPath, "{location}", url.PathEscape(location))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
 	if err != nil {
 		return nil, err
 	}
@@ -95,14 +97,17 @@ func (client *OperationStatusResourceGroupContextClient) getCreateRequest(ctx co
 	reqQP.Set("api-version", "2023-11-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
+		return nil, err
+	}
 	return req, nil
 }
 
 // getHandleResponse handles the Get response.
-func (client *OperationStatusResourceGroupContextClient) getHandleResponse(resp *http.Response) (OperationStatusResourceGroupContextClientGetResponse, error) {
-	result := OperationStatusResourceGroupContextClientGetResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.OperationResource); err != nil {
-		return OperationStatusResourceGroupContextClientGetResponse{}, err
+func (client *FetchCrossRegionRestoreJobClient) getHandleResponse(resp *http.Response) (FetchCrossRegionRestoreJobClientGetResponse, error) {
+	result := FetchCrossRegionRestoreJobClientGetResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.AzureBackupJobResource); err != nil {
+		return FetchCrossRegionRestoreJobClientGetResponse{}, err
 	}
 	return result, nil
 }